requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
fonttools==4.58.4
frozendict==2.4.6
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
narwhals==1.44.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
protobuf==6.31.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.63
streamlit
scipy
